From John Vidar Larring, "While testing OSG 3.0 I also recompiled VPB trunk against it, and came across a few compile warnings from gcc that was easy to fix / remove.
The attached tgz file contains fixes for the following warnings:

* trunk/include/vpb/BuildLog:190: warning: ?class vpb::Logger? has virtual functions but non-virtual destructor
* trunk/include/vpb/Destination:72: warning: ?class vpb::DestinationVisitor? has virtual functions but non-virtual destructor
* trunk/src/vpb/PropertyFile.cpp:321: warning: ?file.vpb::FileProxy::_requiresSync? may be used uninitialized in this function
* trunk/src/vpb/TaskManager.cpp:105: warning: converting to ?int? from ?double?
* trunk/src/vpb/TaskManager.cpp:174: warning: converting to ?int? from ?double?

When patch is applied VPB compiles without warnings."

+++ src/vpb/TaskManager.cpp
@@ -102,7 +102,7 @@ void TaskManager::readPatchSetUp(const std::string& patchFile)
             int revisionNum = -1;
             if (isNumeric)
             {
-                revisionNum = atof(extension.c_str());
+                revisionNum = atoi(extension.c_str());
                 filename = osgDB::getNameLessExtension(filename);
                 extension = osgDB::getFileExtension(filename);
                 basename = filename;
@@ -171,7 +171,7 @@ void TaskManager::readPatchSetUp(const std::string& patchFile)
                     int revisionNum = -1;
                     if (isNumeric)
                     {
-                        revisionNum = atof(revisionExtension.c_str());
+                        revisionNum = atoi(revisionExtension.c_str());
                         sourceMap[revisionNum] = file;
                     }
                     else
